JsonSchema に関する処理を１か所にまとめた
